+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Current Master
 
+- Redesigned init, vault create, vault restore and seed confirmation screens.
 - Added pending transactions to transaction list on account screen.
 - Clicking a pending transaction takes you back to the transaction approval screen.